Remove factorial scaling setup
Currently FQEData which is used by core Wavefunction methods
spends a norb! computation setting up the basis vectors.  The computation of the basis vector strings is changed to to a combination which
avoids performing all norb! permutations of a string and then finding only unique strings.  The previous process was incredibly inefficient.