OpenFermion sync, new notebook
Changes in OpenFermion required changing imports.  normal_ordered changed
location.

I've now set the requirements.txt to temporarily link to the github openfermion.
Right now this is pip failing due to an error on OpenFermion's side (setup.py is
now incorrect because of something unrelated to FQE).

Added a new notebook to show how the diagonal_coulomb evolution works.  This is
a required step in coding up the low-rank decomposition evolution.