Fixed 3-RDM construction
spin-orbital 3-RDM construction now works correctly.

A small bug in the order of alpha-beta dvec multiplications was
the cause for the issues before.  This code has been tested with
open and closed shell states.  I will need to add other wavefuncitons
to the testing suite to directly test this contraction functionality.