Trotterized generalized doubles
previous Low-rank factorization techniques require the amplitudes to have the
standard restriction to occupied and virtual sets.  Here we have implemented
a rederived version for the generalized doubles operator.  The generalized
doubles is the nso**2 limit of k-UpCSGSD wavefunction ansatz.

Here we use it to demonstrate that even with very few singular values
we can accelerate the convergence of ADAPT VQE greatly.

The cost is measuring the full 3-RDM at every step instead of the limited set
of 3-RDM elements required for ADAPT.