pyproject.toml and update mypy
pyproject.toml lets one specify the build requirements for the library.

Cython is required for cythonize.

Numpy and scipy are required for blas_ lapack pointers.

on my linux system typed-ast was having issues with the older version of
mypy (pyObject call issue).  0.9 worked with the current yapf version.

We may have to update the github workflow.  We will see...