Taylor Expansion limit settings
The Taylor expansion and Chebyshev expansion code
had a hard stop at 30 terms and the interface to
`apply_generated_unitary` had a true zero as the
accuracy (which is never achieved in practice).

Adjusting the code such that we no longer require
true zero but instead numerical zero of 1.E-15
revealed a number of problems associated with
taking a very long time and possible issues with
inplace evolution.  I will open issues for these
after this PR.

The algorithms update for Givens rotation and
doubles evolution now no longer uses inplace
evolution which seems to behave correctly.